FormikTapisFile manages its own modal state
