+++ src/tapis-ui/_common/FieldWrapperFormik/fields/index.ts
@@ -10,4 +10,5 @@ export type FormikInputProps = {
 export { default as FormikInput } from './FormikInput';
 export { default as FormikSelect } from './FormikSelect';
 export { default as FormikCheck } from './FormikCheck';
-export { default as FormikBrowseInput } from './FormikBrowseInput';
+export { default as FormikTapisFile } from './FormikTapisFile';
+export { FormikTapisFileInput } from './FormikTapisFile';
