+++ src/tapis-ui/_common/FieldWrapperFormik/index.ts
@@ -3,5 +3,6 @@ export {
   FormikInput,
   FormikSelect,
   FormikCheck,
-  FormikBrowseInput,
+  FormikTapisFile,
+  FormikTapisFileInput,
 } from './fields';
