+++ src/tapis-ui/components/jobs/JobLauncher/steps/FileInputs.tsx
@@ -1,5 +1,5 @@
-import React, { useMemo, useState, useCallback } from 'react';
-import { Apps, Files, Jobs } from '@tapis/tapis-typescript';
+import React, { useMemo } from 'react';
+import { Apps, Jobs } from '@tapis/tapis-typescript';
 import FieldWrapper from 'tapis-ui/_common/FieldWrapper';
 import { Input } from 'reactstrap';
 import { Button } from 'reactstrap';
@@ -12,15 +12,14 @@ import {
   getAppInputsIncludedByDefault,
 } from 'tapis-api/utils/jobFileInputs';
 import { Collapse } from 'tapis-ui/_common';
-import { FieldArray, useFormikContext, FieldArrayRenderProps, useFormik } from 'formik';
+import { FieldArray, useFormikContext, FieldArrayRenderProps } from 'formik';
 import { FormikJobStepWrapper } from '../components';
 import {
   FormikInput,
   FormikCheck,
-  FormikBrowseInput,
+  FormikTapisFile,
 } from 'tapis-ui/_common/FieldWrapperFormik';
 import { v4 as uuidv4 } from 'uuid';
-import { FileSelectModal } from 'tapis-ui/components/files';
 import * as Yup from 'yup';
 
 type FileInputFieldProps = {
@@ -40,7 +39,6 @@ const JobInputField: React.FC<FileInputFieldProps> = ({
   remove,
 }) => {
   const { app } = useJobLauncher();
-  const { setFieldValue } = useFormikContext();
   const { name, sourceUrl } = item;
   const inputMode: Apps.FileInputModeEnum | undefined = useMemo(
     () =>
@@ -54,19 +52,7 @@ const JobInputField: React.FC<FileInputFieldProps> = ({
   const note = `${
     inputMode ? upperCaseFirstLetter(inputMode) : 'User Defined'
   }`;
-  const [modalOpen, setModalOpen] = useState(false);
-  const onBrowse = useCallback(() => {
-    setModalOpen(true);
-  }, [setModalOpen]);
-  const toggle = useCallback(() => {
-    setModalOpen(false);
-  }, [setModalOpen]);
-  const onSelect = useCallback(
-    (systemId: string | null, files: Array<Files.FileInfo>) => {
-      setFieldValue(`fileInputs.${index}.sourceUrl`, `tapis://${systemId ?? ''}${files[0].path}`);
-    },
-    [setFieldValue]
-  )
+
   return (
     <>
       <Collapse
@@ -87,12 +73,11 @@ const JobInputField: React.FC<FileInputFieldProps> = ({
           }`}
           disabled={isRequired}
         />
-        <FormikBrowseInput
+        <FormikTapisFile
           name={`fileInputs.${index}.sourceUrl`}
           label="Source URL"
           required={true}
           description="Input TAPIS file as a pathname, TAPIS URI or web URL"
-          onBrowse={onBrowse}
         />
         <FormikInput
           name={`fileInputs.${index}.targetPath`}
@@ -118,13 +103,6 @@ const JobInputField: React.FC<FileInputFieldProps> = ({
           </Button>
         )}
       </Collapse>
-      {modalOpen && (
-        <FileSelectModal
-          toggle={toggle} 
-          selectMode={{mode: 'single', types: ['dir', 'file']}}
-          onSelect={onSelect}
-        />
-      )}
     </>
   );
 };
